+++ tests/primary_query_test.py
@@ -130,6 +130,7 @@ def test_sense_empty():
     with pytest.raises(wn.Error):
         assert wn.sense('test-es-información-n-0001-01')
 
+
 @pytest.mark.usefixtures('mini_db')
 def test_sense_mini():
     assert wn.sense('test-es-información-n-0001-01')
@@ -140,3 +141,60 @@ def test_sense_mini():
     with pytest.raises(wn.Error):
         assert wn.sense('test-es-información-n-0001-01', lexicon='test-en')
 
+
+@pytest.mark.usefixtures('empty_db')
+def test_synsets_empty():
+    assert len(wn.synsets()) == 0
+
+
+@pytest.mark.usefixtures('mini_db')
+def test_synsets_mini():
+    assert len(wn.synsets()) == 6
+    assert all(isinstance(ss, wn.Synset) for ss in wn.synsets())
+
+    synsets = wn.synsets('information')  # search lemma
+    assert len(synsets) == 1
+    assert 'information' in synsets[0].lemmas()
+
+    synsets = wn.synsets('exemplifies')  # search secondary form
+    assert len(synsets) == 1
+    assert 'exemplify' in synsets[0].lemmas()
+
+    assert len(wn.synsets(pos='n')) == 4
+    assert len(wn.synsets(pos='v')) == 2
+    assert len(wn.synsets(pos='q')) == 0  # fake pos
+
+    assert len(wn.synsets(ili='i67469')) == 2
+    assert len(wn.synsets(ili='i67468')) == 0
+
+    assert len(wn.synsets(lgcode='en')) == 3
+    assert len(wn.synsets(lgcode='es')) == 3
+    # assert len(wn.synsets(lgcode='pt')) == 0
+
+    assert len(wn.synsets(lexicon='test-en')) == 3
+    assert len(wn.synsets(lexicon='test-es')) == 3
+    # assert len(wn.synsets(lexicon='test-pt')) == 0
+
+    assert len(wn.synsets(lgcode='en', lexicon='test-en')) == 3
+    # assert len(wn.synsets(lgcode='en', lexicon='test-es')) == 0
+    assert len(wn.synsets(pos='v', lgcode='en')) == 1
+    assert len(wn.synsets('information', lgcode='en')) == 1
+    assert len(wn.synsets('information', lgcode='es')) == 0
+    assert len(wn.synsets(ili='i67469', lgcode='es')) == 1
+
+
+@pytest.mark.usefixtures('empty_db')
+def test_synset_empty():
+    with pytest.raises(wn.Error):
+        assert wn.synset('test-es-0001-n')
+
+
+@pytest.mark.usefixtures('mini_db')
+def test_synset_mini():
+    assert wn.synset('test-es-0001-n')
+    assert wn.synset('test-es-0001-n', lgcode='es')
+    assert wn.synset('test-es-0001-n', lexicon='test-es')
+    with pytest.raises(wn.Error):
+        assert wn.synset('test-es-0001-n', lgcode='en')
+    with pytest.raises(wn.Error):
+        assert wn.synset('test-es-0001-n', lexicon='test-en')